- moved actor_crtics specific functions into algorithm to allow generalizability of learning_role functionalities -fix continue learning with pretrained strategies , since learning strategies were loaded before units were added, which leads to not loading anything
